requirements.txt
click>=8.0.0
requests>=2.25.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
pyyaml>=5.4.0
rich>=10.0.0
tabulate>=0.8.9
kubernetes>=18.0.0
python-dateutil>=2.8.0
colorama>=0.4.4
tqdm>=4.62.0 